# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Delivery Service API v2
  description: |
    물품 배송 서비스 백엔드 API 명세서

    ## Architecture
    - Hexagonal Architecture (Ports & Adapters)
    - Spring Boot 4.0.1 / Java 21
    - Event-Driven Architecture

    ## Service Type
    - Product Delivery Service (물품 배송 서비스)
    - Courier-based delivery (택배사 연동)
    - Domestic shipping only (국내 배송)
    - Return/Exchange supported (반품/교환 지원)

    ## Domains
    1. Customer - 고객 관리
    2. Product - 상품 관리
    3. Seller - 판매자 관리
    4. Cart - 장바구니
    5. Order - 주문 관리
    6. Payment - 결제 관리
    7. Shipment - 배송 관리
    8. Return - 반품/교환 관리
    9. Review - 리뷰 관리
    10. Promotion - 쿠폰/프로모션
  version: 2.0.0
  contact:
    name: Delivery Service Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v2
    description: Local Development Server
  - url: https://api.delivery.com/api/v2
    description: Production Server
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: 상품 목록 조회
      operationId: getProducts
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
        - name: keyword
          in: query
          schema:
            type: string
        - name: sellerId
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [recent, price_asc, price_desc, rating, sales]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /products/{productId}:
    get:
      tags: [Products]
      summary: 상품 상세 조회
      operationId: getProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /products/{productId}/reviews:
    get:
      tags: [Products]
      summary: 상품 리뷰 목록
      operationId: getProductReviews
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - name: sort
          in: query
          schema:
            type: string
            enum: [recent, rating_high, rating_low]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'

  # ============================================
  # Seller API (Public)
  # ============================================
...
  /reviews:
    post:
      tags: [Reviews]
      summary: 리뷰 작성
      operationId: createReview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: 작성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
  /reviews/{reviewId}:
    put:
      tags: [Reviews]
      summary: 리뷰 수정
      operationId: updateReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200':
          description: 수정 성공

    delete:
      tags: [Reviews]
      summary: 리뷰 삭제
      operationId: deleteReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '204':
          description: 삭제 성공

  # ============================================
  # Coupon API
  # ============================================
...
  /sellers/me/products/{productId}:
    put:
      tags: [Seller Admin]
      summary: 상품 수정
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '200':
          description: 수정 성공

    delete:
      tags: [Seller Admin]
      summary: 상품 삭제
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: 삭제 성공
...
  /sellers/me/reviews/{reviewId}/reply:
    post:
      tags: [Seller Admin]
      summary: 리뷰 답글 작성
      operationId: createReviewReply
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReplyRequest'
      responses:
        '201':
          description: 작성 성공

    put:
      tags: [Seller Admin]
      summary: 리뷰 답글 수정
      operationId: updateReviewReply
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReplyRequest'
      responses:
        '200':
          description: 수정 성공

    delete:
      tags: [Seller Admin]
      summary: 리뷰 답글 삭제
      operationId: deleteReviewReply
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '204':
          description: 삭제 성공

  # ============================================
  # Webhooks
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        default: 20
...
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
...
    ReviewId:
      name: reviewId
      in: path
      required: true
      schema:
        type: string
  responses:
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ProductStatus:
      type: string
      enum:
        - DRAFT
        - ACTIVE
        - OUT_OF_STOCK
        - INACTIVE
        - DELETED
...
    CreateReviewRequest:
      type: object
      required: [orderId, productId, rating]
      properties:
        orderId:
          type: string
        productId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        content:
          type: string
        images:
          type: array
          items:
            type: string
    UpdateReviewRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        content:
          type: string
        images:
          type: array
          items:
            type: string
    CreateProductRequest:
      type: object
      required: [categoryId, name, price]
      properties:
        categoryId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
        sku:
          type: string
        images:
          type: array
          items:
            type: object
            properties:
              imageUrl:
                type: string
              isPrimary:
                type: boolean
        variants:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              sku:
                type: string
              price:
                type: number
              stock:
                type: integer
              optionValues:
                type: object
                additionalProperties:
                  type: string
...
    CreateReplyRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ============================================
    # Response Schemas
    # ============================================
...
    ProductListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummaryResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    ProductSummaryResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        thumbnailUrl:
          type: string
        ratingAvg:
          type: number
        reviewCount:
          type: integer
        sellerId:
          type: string
        sellerName:
          type: string
    ProductDetailResponse:
      type: object
      properties:
        id:
          type: string
        sellerId:
          type: string
        sellerName:
          type: string
        categoryId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        stock:
          type: integer
        sku:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        isFeatured:
          type: boolean
        images:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              imageUrl:
                type: string
              isPrimary:
                type: boolean
        variants:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              sku:
                type: string
              price:
                type: number
              stock:
                type: integer
              optionValues:
                type: object
                additionalProperties:
                  type: string
        ratingAvg:
          type: number
        reviewCount:
          type: integer
        salesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    ReviewResponse:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        productId:
          type: string
        rating:
          type: integer
        content:
          type: string
        images:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              imageUrl:
                type: string
        reply:
          type: object
          properties:
            content:
              type: string
            createdAt:
              type: string
              format: date-time
        createdAt:
          type: string
          format: date-time
    ReviewListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ReviewSummaryResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        ratingDistribution:
          type: object
          additionalProperties:
            type: integer
    ReviewSummaryResponse:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        customerName:
          type: string
        rating:
          type: integer
        content:
          type: string
        images:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              imageUrl:
                type: string
        reply:
          type: object
          properties:
            content:
              type: string
            createdAt:
              type: string
              format: date-time
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
        timestamp:
          type: string
          format: date-time